Implement full linked-list data storage (#2)
Implement an in-memory linked list data storage engine for both adding and querying points based on their timestamp. This helps cement the PointIO interface that can be used by all data storage engines. Included in this is a generic test harness using this interface that can be used for other data storage engine. Included in this is a first implementation the Query/QueryExec/Cursor interfaces for retrieving data in batches. Querying on attributes is not implemented yet.