Add unique identifier to Point struct (#7)
Added a unique identifier to all new points that get created and an Identical() function to help with identity checks. Start using this Id to ensure strict ordering within MemTree as required by the BTRee implementation. For now the Id is a uint64 that can be converted to/from base64 encoding for use in URLs. Made some minor test improvements along the way.