Add REST endpoint for adding a point (#10)
Add the REST endpoint /series/{id}/points for adding points to a data series along with unit tests. Includes several changes to support JSON marshaling and unmarshaling of data such as Id and Point. Responses are sent in JSend-compliant format, and I added a class to make that easier to manage. I added an initial/placeholder Series model struct and SeriesMgr singleton struct to manage access to it.